fix(ui): cast HistoricalProbeKind to u8 in sort_by_key (enum lacks Ord)
HistoricalProbeKind derives Component, Debug, Clone, Copy, PartialEq, Eq,
Hash but not PartialOrd/Ord, so `*kind` in `sort_by_key` is a type error.
The linter didn't catch it because the previous commit (e122e81) ran with
the worktree's edited version, not the committed version — CI builds from
the committed tree and saw the compilation error in the `cargo clippy`
step (run 27646675826, completed in 14s — consistent with a compile
failure, not a lint).

Cast the enum discriminant via `*kind as u8` so the sort still works on
the declaration order (Voyager1 < Voyager2 < Parker < NewHorizons).  The
cast is on a derived-enum without #[repr(u8)] — clippy::cast_enum_to_primitive
is restriction-level and doesn't trip -D warnings.

+++ src/ui/fleets_panel.rs
@@ -937,9 +937,12 @@ fn render_historical_probes_section(
             )
         })
         .collect();
-    // HistoricalProbeKind is `#[derive(PartialOrd, Ord)]` over its
-    // declaration order (Voyager1 < Voyager2 < Parker < NewHorizons).
-    rows.sort_by_key(|(kind, _)| *kind);
+    // HistoricalProbeKind is declared in display order
+    // (Voyager1 → Voyager2 → Parker → NewHorizons); sort by the
+    // discriminant so the table is stable across runs.  We use
+    // `as u8` because the enum derives `Eq`/`Hash`/`Copy` but
+    // not `Ord` — see `src/fleets/components.rs:733`.
+    rows.sort_by_key(|(kind, _)| *kind as u8);
 
     ui.label(
         egui::RichText::new("DEEP SPACE PROBES")
